Module grpc_unary_attestation::proto::unary_session_server

Generated server implementations.

Structs

UnarySessionServer

Service definition for unary communication with the runtime. It accomplishes the same goal as StreamingSession, just with a differnet request model. Oak Applications can choose either request model in the configuration of the runtime. Eventually we plan to transition all clients to use UnarySessions.

Traits

UnarySession

Generated trait containing gRPC methods that should be implemented for use with UnarySessionServer.