Trait grpc_unary_attestation::proto::unary_session_server::UnarySession

pub trait UnarySession: Send + Sync + 'static {
    fn message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UnaryRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<UnaryResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with UnarySessionServer.

Required Methods

fn message<'life0, 'async_trait>(
&'life0 self,
request: Request<UnaryRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<UnaryResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Used to send a sequence of a messages for session key negotiation and encrypted payload exchange.

The expected message seqeuence starts with an intial handshake:

Client->Server: UnaryRequest with a serialized ClientHello message.

Server->Client: UnaryResponse with a serialized ServerIdentity message.

Client->Server: UnaryRequest with a serialized ClientIdentity message.

Server->Client: UnaryResponse with an empty message, confirming handshake completion.

After the handshake, both client and server derive matching session keys and are then able to exchange multiple UnaryRequest/UnaryResponse request pairs that contain a seralized EncryptedData message:

Client->Server: UnaryRequest with a serialized EncryptedData message.

Server->Client: UnaryResponse with a serialized EncryptedData message.

Messages are represented as serialized messages defined in the remote_attestation::message.rs and com.google.oak.remote_attestation.Message.